requirements.txt
Flask==0.10.1
imutils==0.5.3
opencv-python-headless
requests
Pillow